+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 - Chg #225: Rename classes: `All` to `AndX`, `Any` to `OrX`. Remove `Group` class (@vjik)
 - Chg #226: Refactor filter classes to use readonly properties instead of getters (@vjik)
 - New #213: Add `nextPage()` and `previousPage()` methods to `PaginatorInterface` (@samdark)
+- New #200: Add matching mode parameter to `Like` filter (@samdark, @vjik)
 
 ## 1.0.1 January 25, 2023